refactor: [app] move getUserProfile plugin
